rename: 이미지 저장소 엔티티 파일 이동
