+++ src/main/java/project/volunteer/domain/image/dao/StorageRepository.java
@@ -1,7 +1,7 @@
-package project.volunteer.domain.storage.dao;
+package project.volunteer.domain.image.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import project.volunteer.domain.storage.domain.Storage;
+import project.volunteer.domain.image.domain.Storage;
 
 public interface StorageRepository extends JpaRepository<Storage, Long> {
 }
